+++ crates/meilisearch/src/option.rs
@@ -436,7 +436,7 @@ pub struct Opt {
     /// see: <https://github.com/orgs/meilisearch/discussions/801>
     #[clap(long, env = MEILI_EXPERIMENTAL_LIMIT_BATCHED_TASKS_TOTAL_SIZE, default_value_t = default_limit_batched_tasks_total_size())]
     #[serde(default = "default_limit_batched_tasks_total_size")]
-    pub experimental_limit_batched_tasks_total_size: usize,
+    pub experimental_limit_batched_tasks_total_size: u64,
 
     #[serde(flatten)]
     #[clap(flatten)]
@@ -931,8 +931,8 @@ fn default_limit_batched_tasks() -> usize {
     usize::MAX
 }
 
-fn default_limit_batched_tasks_total_size() -> usize {
-    usize::MAX
+fn default_limit_batched_tasks_total_size() -> u64 {
+    u64::MAX
 }
 
 fn default_snapshot_dir() -> PathBuf {
